moved index to lib and other small cleanup items
