+++ package.json
@@ -1,11 +1,11 @@
 {
   "name": "hueniversity",
   "version": "0.0.1",
-  "description": "This is a community learning experiment. The idea is simple - use GitHub as a platform for teaching people coding skills as a group, where everyone is both a student and a teacher. The goal is to learn how to operate such a distributed classroom and then apply that pattern to other topics by other people.",
-  "main": "index.js",
+  "description": "",
+  "main": "lib/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node index.js"
+    "start": "node lib/index.js"
   },
   "repository": {
     "type": "git",
